+++ feed-crawler/src/feed-crawler.ts
@@ -3,22 +3,20 @@ import { RssRepository } from "./repository/rss.repository";
 import logger from "./common/logger";
 import { RssObj, FeedDetail, RawFeed } from "./common/types";
 import { XMLParser } from "fast-xml-parser";
-import { parse } from "node-html-parser";
-import { unescape } from "html-escaper";
-import { ONE_MINUTE, INTERVAL } from "./common/constant";
+import { ONE_MINUTE, TIME_INTERVAL } from "./common/constant";
+import { RssParser } from "./common/rss-parser";
 
 export class FeedCrawler {
-  private rssParser: RssParser = new RssParser();
   constructor(
     private readonly rssRepository: RssRepository,
     private readonly feedRepository: FeedRepository,
+    private readonly rssParser: RssParser
   ) {}
 
   async start() {
     await this.feedRepository.deleteRecentFeed();
 
     const rssObjects = await this.rssRepository.selectAllRss();
-
     if (!rssObjects || !rssObjects.length) {
       logger.info("등록된 RSS가 없습니다.");
       return;
@@ -32,7 +30,6 @@ export class FeedCrawler {
       return;
     }
     logger.info(`총 ${newFeeds.length}개의 새로운 피드가 있습니다.`);
-
     const insertedData = await this.feedRepository.insertFeeds(newFeeds);
 
     await this.feedRepository.setRecentFeedList(insertedData);
@@ -40,12 +37,10 @@ export class FeedCrawler {
 
   private async findNewFeeds(
     rssObj: RssObj,
-    now: number,
+    now: number
   ): Promise<FeedDetail[]> {
     try {
-      const TIME_INTERVAL = INTERVAL;
       const feeds = await this.fetchRss(rssObj.rssUrl);
-
       const filteredFeeds = feeds.filter((item) => {
         const pubDate = new Date(item.pubDate).setSeconds(0, 0);
         const timeDiff = (now - pubDate) / (ONE_MINUTE * TIME_INTERVAL);
@@ -71,13 +66,13 @@ export class FeedCrawler {
             link: decodeURIComponent(feed.link),
             imageUrl: imageUrl,
           };
-        }),
+        })
       );
 
       return detailedFeeds;
     } catch (err) {
       logger.warn(
-        `[${rssObj.rssUrl}] 에서 데이터 조회 중 오류 발생으로 인한 스킵 처리. 오류 내용 : ${err}`,
+        `[${rssObj.rssUrl}] 에서 데이터 조회 중 오류 발생으로 인한 스킵 처리. 오류 내용 : ${err}`
       );
       return [];
     }
@@ -87,10 +82,10 @@ export class FeedCrawler {
     return Promise.all(
       rssObjects.map(async (rssObj: RssObj) => {
         logger.info(
-          `${rssObj.blogName}(${rssObj.rssUrl}) 에서 데이터 조회하는 중...`,
+          `${rssObj.blogName}(${rssObj.rssUrl}) 에서 데이터 조회하는 중...`
         );
         return await this.findNewFeeds(rssObj, currentTime.setSeconds(0, 0));
-      }),
+      })
     );
   }
 
@@ -117,59 +112,5 @@ export class FeedCrawler {
       link: feed.link,
       pubDate: feed.pubDate,
     }));
-  }
-}
-
-class RssParser {
-  async getThumbnailUrl(feedUrl: string) {
-    const response = await fetch(feedUrl, {
-      headers: {
-        Accept: "text/html",
-      },
-    });
-
-    if (!response.ok) {
-      throw new Error(`${feedUrl}에 GET 요청 실패`);
-    }
-
-    const htmlData = await response.text();
-    const htmlRootElement = parse(htmlData);
-    const metaImage = htmlRootElement.querySelector(
-      'meta[property="og:image"]',
-    );
-    let thumbnailUrl = metaImage?.getAttribute("content") ?? "";
-
-    if (!thumbnailUrl.length) {
-      logger.warn(`${feedUrl}에서 썸네일 추출 실패`);
-      return thumbnailUrl;
-    }
-
-    if (!this.isUrlPath(thumbnailUrl)) {
-      thumbnailUrl = this.getHttpOriginPath(feedUrl) + thumbnailUrl;
-    }
-    return thumbnailUrl;
-  }
-
-  private isUrlPath(thumbnailUrl: string) {
-    const reg = /^(http|https):\/\//;
-    return reg.test(thumbnailUrl);
-  }
-
-  private getHttpOriginPath(feedUrl: string) {
-    return new URL(feedUrl).origin;
-  }
-
-  customUnescape(feedTitle: string): string {
-    const escapeEntity = {
-      "&middot;": "·",
-      "&nbsp;": " ",
-    };
-    Object.keys(escapeEntity).forEach((escapeKey) => {
-      const value = escapeEntity[escapeKey];
-      const regex = new RegExp(escapeKey, "g");
-      feedTitle = feedTitle.replace(regex, value);
-    });
-
-    return unescape(feedTitle);
   }
 }
