+++ feed-crawler/src/main.ts
@@ -1,4 +1,5 @@
 import "reflect-metadata";
+import "./common/env-load";
 import logger from "./common/logger";
 import { FeedCrawler } from "./feed-crawler";
 import { container } from "./container";
@@ -6,6 +7,7 @@ import { RssRepository } from "./repository/rss.repository";
 import { FeedRepository } from "./repository/feed.repository";
 import { DEPENDENCY_SYMBOLS } from "./types/dependency-symbols";
 import { DatabaseConnection } from "./types/database-connection";
+import { RssParser } from "./common/rss-parser";
 
 async function main() {
   logger.info("==========작업 시작==========");
@@ -20,8 +22,9 @@ async function main() {
   const dbConnection = container.resolve<DatabaseConnection>(
     DEPENDENCY_SYMBOLS.DatabaseConnection
   );
+  const rssParser = container.resolve<RssParser>(DEPENDENCY_SYMBOLS.RssParser);
 
-  const feedCrawler = new FeedCrawler(rssRepository, feedRepository);
+  const feedCrawler = new FeedCrawler(rssRepository, feedRepository, rssParser);
   await feedCrawler.start();
 
   const endTime = Date.now();
