+++ feed-crawler/src/types/dependency-symbols.ts
@@ -1,6 +1,7 @@
 export const DEPENDENCY_SYMBOLS = {
-    DatabaseConnection: Symbol.for("DatabaseConnection"),
-    RssRepository: Symbol.for("RssRepository"),
-    FeedRepository: Symbol.for("FeedRepository"),
-    RedisConnection: Symbol.for("RedisConnection"),
-}
+  DatabaseConnection: Symbol.for("DatabaseConnection"),
+  RssRepository: Symbol.for("RssRepository"),
+  FeedRepository: Symbol.for("FeedRepository"),
+  RedisConnection: Symbol.for("RedisConnection"),
+  RssParser: Symbol.for("RssParser"),
+};
